requirements.txt
# Test dependencies
pytest>=7.0
pytest-cov>=4.0
pyyaml>=6.0
